recording-booth/backend/services/qr_gen.py
import io
import qrcode
from PIL import Image


def generate_qr_bytes(url: str, size: int = 400) -> bytes:
    """Generate a QR code PNG for the given URL. Returns raw PNG bytes."""
    qr = qrcode.QRCode(
        version=None,
        error_correction=qrcode.constants.ERROR_CORRECT_M,
        box_size=10,
        border=4,
    )
    qr.add_data(url)
    qr.make(fit=True)

    img = qr.make_image(fill_color="black", back_color="white").convert("RGB")
    img = img.resize((size, size), Image.LANCZOS)

    buf = io.BytesIO()
    img.save(buf, format="PNG")
    buf.seek(0)
    return buf.getvalue()
